LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text-styles/2d.text.draw.space.collapse.end-expected.txt

2d.text.draw.space.collapse.end
Space characters at the end of a line are collapsed (per CSS)
Actual output:
Expected output:


FAIL Space characters at the end of a line are collapsed (per CSS) assert_approx_equals: Red channel of the pixel at (75, 25) expected 0 +/- 2 but got 255
